Revert "Refactor: Migrate to 2.0-style security policies (#11218)"
This reverts commit 975a485a96c83644221cdf626706c2be15b8a089.
