+++ tests/unit/macaroons/test_auth_policy.py
@@ -10,15 +10,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import uuid
 
 import pretend
 import pytest
 
-from pyramid.interfaces import IAuthorizationPolicy, ISecurityPolicy
+from pyramid.interfaces import IAuthenticationPolicy, IAuthorizationPolicy
 from pyramid.security import Denied
 from zope.interface.verify import verifyClass
 
-from warehouse.macaroons import security_policy
+from warehouse.macaroons import auth_policy
 from warehouse.macaroons.interfaces import IMacaroonService
 from warehouse.macaroons.services import InvalidMacaroonError
 
@@ -38,7 +39,7 @@ def test_extract_http_macaroon(auth, result):
         headers=pretend.stub(get=pretend.call_recorder(lambda k: auth))
     )
 
-    assert security_policy._extract_http_macaroon(request) == result
+    assert auth_policy._extract_http_macaroon(request) == result
 
 
 @pytest.mark.parametrize(
@@ -52,99 +53,123 @@ def test_extract_http_macaroon(auth, result):
     ],
 )
 def test_extract_basic_macaroon(auth, result):
-    assert security_policy._extract_basic_macaroon(auth) == result
+    assert auth_policy._extract_basic_macaroon(auth) == result
 
 
-class TestMacaroonSecurityPolicy:
+class TestMacaroonAuthenticationPolicy:
     def test_verify(self):
         assert verifyClass(
-            ISecurityPolicy,
-            security_policy.MacaroonSecurityPolicy,
+            IAuthenticationPolicy, auth_policy.MacaroonAuthenticationPolicy
         )
 
-    def test_noops(self):
-        policy = security_policy.MacaroonSecurityPolicy()
-        assert policy.authenticated_userid(pretend.stub()) == NotImplemented
-        assert (
-            policy.permits(pretend.stub(), pretend.stub(), pretend.stub())
-            == NotImplemented
+    def test_unauthenticated_userid_invalid_macaroon(self, monkeypatch):
+        _extract_http_macaroon = pretend.call_recorder(lambda r: None)
+        monkeypatch.setattr(
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
         )
 
-    def test_forget_and_remember(self):
-        policy = security_policy.MacaroonSecurityPolicy()
-
-        assert policy.forget(pretend.stub()) == []
-        assert policy.remember(pretend.stub(), pretend.stub()) == []
-
-    def test_identify_no_macaroon(self, monkeypatch):
-        policy = security_policy.MacaroonSecurityPolicy()
+        policy = auth_policy.MacaroonAuthenticationPolicy()
 
         vary_cb = pretend.stub()
         add_vary_cb = pretend.call_recorder(lambda *v: vary_cb)
-        monkeypatch.setattr(security_policy, "add_vary_callback", add_vary_cb)
-
-        extract_http_macaroon = pretend.call_recorder(lambda r: None)
-        monkeypatch.setattr(
-            security_policy, "_extract_http_macaroon", extract_http_macaroon
-        )
+        monkeypatch.setattr(auth_policy, "add_vary_callback", add_vary_cb)
 
         request = pretend.stub(
             add_response_callback=pretend.call_recorder(lambda cb: None)
         )
 
-        assert policy.identity(request) is None
-        assert extract_http_macaroon.calls == [pretend.call(request)]
-
+        assert policy.unauthenticated_userid(request) is None
+        assert _extract_http_macaroon.calls == [pretend.call(request)]
         assert add_vary_cb.calls == [pretend.call("Authorization")]
         assert request.add_response_callback.calls == [pretend.call(vary_cb)]
 
-    def test_identify(self, monkeypatch):
-        policy = security_policy.MacaroonSecurityPolicy()
+    def test_unauthenticated_userid_valid_macaroon(self, monkeypatch):
+        _extract_http_macaroon = pretend.call_recorder(lambda r: b"not a real macaroon")
+        monkeypatch.setattr(
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
+        )
+
+        policy = auth_policy.MacaroonAuthenticationPolicy()
 
         vary_cb = pretend.stub()
         add_vary_cb = pretend.call_recorder(lambda *v: vary_cb)
-        monkeypatch.setattr(security_policy, "add_vary_callback", add_vary_cb)
+        monkeypatch.setattr(auth_policy, "add_vary_callback", add_vary_cb)
 
-        raw_macaroon = pretend.stub()
-        extract_http_macaroon = pretend.call_recorder(lambda r: raw_macaroon)
-        monkeypatch.setattr(
-            security_policy, "_extract_http_macaroon", extract_http_macaroon
-        )
-
-        user = pretend.stub()
+        userid = uuid.uuid4()
         macaroon_service = pretend.stub(
-            find_from_raw=pretend.call_recorder(lambda m: pretend.stub(user=user))
+            find_userid=pretend.call_recorder(lambda macaroon: userid)
         )
         request = pretend.stub(
+            find_service=pretend.call_recorder(
+                lambda interface, **kw: macaroon_service
+            ),
             add_response_callback=pretend.call_recorder(lambda cb: None),
-            find_service=pretend.call_recorder(lambda i, **kw: macaroon_service),
         )
 
-        assert policy.identity(request) is user
-        assert extract_http_macaroon.calls == [pretend.call(request)]
+        assert policy.unauthenticated_userid(request) == str(userid)
+        assert _extract_http_macaroon.calls == [pretend.call(request)]
         assert request.find_service.calls == [
             pretend.call(IMacaroonService, context=None)
         ]
-        assert macaroon_service.find_from_raw.calls == [pretend.call(raw_macaroon)]
-
+        assert macaroon_service.find_userid.calls == [
+            pretend.call(b"not a real macaroon")
+        ]
         assert add_vary_cb.calls == [pretend.call("Authorization")]
         assert request.add_response_callback.calls == [pretend.call(vary_cb)]
+
+    def test_unauthenticated_userid_valid_macaroon_invalid_userid(self, monkeypatch):
+        _extract_http_macaroon = pretend.call_recorder(lambda r: b"not a real macaroon")
+        monkeypatch.setattr(
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
+        )
+
+        policy = auth_policy.MacaroonAuthenticationPolicy()
+
+        vary_cb = pretend.stub()
+        add_vary_cb = pretend.call_recorder(lambda *v: vary_cb)
+        monkeypatch.setattr(auth_policy, "add_vary_callback", add_vary_cb)
+
+        macaroon_service = pretend.stub(
+            find_userid=pretend.call_recorder(lambda macaroon: None)
+        )
+        request = pretend.stub(
+            find_service=pretend.call_recorder(
+                lambda interface, **kw: macaroon_service
+            ),
+            add_response_callback=pretend.call_recorder(lambda cb: None),
+        )
+
+        assert policy.unauthenticated_userid(request) is None
+        assert _extract_http_macaroon.calls == [pretend.call(request)]
+        assert add_vary_cb.calls == [pretend.call("Authorization")]
+        assert macaroon_service.find_userid.calls == [
+            pretend.call(b"not a real macaroon")
+        ]
+        assert request.add_response_callback.calls == [pretend.call(vary_cb)]
+
+    def test_remember(self):
+        policy = auth_policy.MacaroonAuthenticationPolicy()
+        assert policy.remember(pretend.stub(), pretend.stub()) == []
+
+    def test_forget(self):
+        policy = auth_policy.MacaroonAuthenticationPolicy()
+        assert policy.forget(pretend.stub()) == []
 
 
 class TestMacaroonAuthorizationPolicy:
     def test_verify(self):
         assert verifyClass(
-            IAuthorizationPolicy, security_policy.MacaroonAuthorizationPolicy
+            IAuthorizationPolicy, auth_policy.MacaroonAuthorizationPolicy
         )
 
     def test_permits_no_active_request(self, monkeypatch):
         get_current_request = pretend.call_recorder(lambda: None)
-        monkeypatch.setattr(security_policy, "get_current_request", get_current_request)
+        monkeypatch.setattr(auth_policy, "get_current_request", get_current_request)
 
         backing_policy = pretend.stub(
             permits=pretend.call_recorder(lambda *a, **kw: pretend.stub())
         )
-        policy = security_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
+        policy = auth_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
         result = policy.permits(pretend.stub(), pretend.stub(), pretend.stub())
 
         assert result == Denied("")
@@ -153,18 +178,18 @@ class TestMacaroonAuthorizationPolicy:
     def test_permits_no_macaroon(self, monkeypatch):
         request = pretend.stub()
         get_current_request = pretend.call_recorder(lambda: request)
-        monkeypatch.setattr(security_policy, "get_current_request", get_current_request)
+        monkeypatch.setattr(auth_policy, "get_current_request", get_current_request)
 
         _extract_http_macaroon = pretend.call_recorder(lambda r: None)
         monkeypatch.setattr(
-            security_policy, "_extract_http_macaroon", _extract_http_macaroon
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
         )
 
         permits = pretend.stub()
         backing_policy = pretend.stub(
             permits=pretend.call_recorder(lambda *a, **kw: permits)
         )
-        policy = security_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
+        policy = auth_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
         result = policy.permits(pretend.stub(), pretend.stub(), pretend.stub())
 
         assert result == permits
@@ -177,18 +202,18 @@ class TestMacaroonAuthorizationPolicy:
             find_service=pretend.call_recorder(lambda interface, **kw: macaroon_service)
         )
         get_current_request = pretend.call_recorder(lambda: request)
-        monkeypatch.setattr(security_policy, "get_current_request", get_current_request)
+        monkeypatch.setattr(auth_policy, "get_current_request", get_current_request)
 
         _extract_http_macaroon = pretend.call_recorder(lambda r: b"not a real macaroon")
         monkeypatch.setattr(
-            security_policy, "_extract_http_macaroon", _extract_http_macaroon
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
         )
 
         permits = pretend.stub()
         backing_policy = pretend.stub(
             permits=pretend.call_recorder(lambda *a, **kw: permits)
         )
-        policy = security_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
+        policy = auth_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
         result = policy.permits(pretend.stub(), pretend.stub(), pretend.stub())
 
         assert result == Denied("")
@@ -202,18 +227,18 @@ class TestMacaroonAuthorizationPolicy:
             find_service=pretend.call_recorder(lambda interface, **kw: macaroon_service)
         )
         get_current_request = pretend.call_recorder(lambda: request)
-        monkeypatch.setattr(security_policy, "get_current_request", get_current_request)
+        monkeypatch.setattr(auth_policy, "get_current_request", get_current_request)
 
         _extract_http_macaroon = pretend.call_recorder(lambda r: b"not a real macaroon")
         monkeypatch.setattr(
-            security_policy, "_extract_http_macaroon", _extract_http_macaroon
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
         )
 
         permits = pretend.stub()
         backing_policy = pretend.stub(
             permits=pretend.call_recorder(lambda *a, **kw: permits)
         )
-        policy = security_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
+        policy = auth_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
         result = policy.permits(pretend.stub(), pretend.stub(), "upload")
 
         assert result == permits
@@ -232,18 +257,18 @@ class TestMacaroonAuthorizationPolicy:
             find_service=pretend.call_recorder(lambda interface, **kw: macaroon_service)
         )
         get_current_request = pretend.call_recorder(lambda: request)
-        monkeypatch.setattr(security_policy, "get_current_request", get_current_request)
+        monkeypatch.setattr(auth_policy, "get_current_request", get_current_request)
 
         _extract_http_macaroon = pretend.call_recorder(lambda r: b"not a real macaroon")
         monkeypatch.setattr(
-            security_policy, "_extract_http_macaroon", _extract_http_macaroon
+            auth_policy, "_extract_http_macaroon", _extract_http_macaroon
         )
 
         permits = pretend.stub()
         backing_policy = pretend.stub(
             permits=pretend.call_recorder(lambda *a, **kw: permits)
         )
-        policy = security_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
+        policy = auth_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
         result = policy.permits(pretend.stub(), pretend.stub(), invalid_permission)
 
         assert result == Denied("")
@@ -258,7 +283,7 @@ class TestMacaroonAuthorizationPolicy:
                 lambda *a: principals
             )
         )
-        policy = security_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
+        policy = auth_policy.MacaroonAuthorizationPolicy(policy=backing_policy)
 
         assert (
             policy.principals_allowed_by_permission(pretend.stub(), pretend.stub())
