+++ warehouse/macaroons/auth_policy.py
@@ -12,7 +12,8 @@
 
 import base64
 
-from pyramid.interfaces import IAuthorizationPolicy, ISecurityPolicy
+from pyramid.authentication import CallbackAuthenticationPolicy
+from pyramid.interfaces import IAuthenticationPolicy, IAuthorizationPolicy
 from pyramid.threadlocal import get_current_request
 from zope.interface import implementer
 
@@ -20,7 +21,6 @@ from warehouse.cache.http import add_vary_callback
 from warehouse.errors import WarehouseDenied
 from warehouse.macaroons.interfaces import IMacaroonService
 from warehouse.macaroons.services import InvalidMacaroonError
-from warehouse.utils.security_policy import AuthenticationMethod
 
 
 def _extract_basic_macaroon(auth):
@@ -70,14 +70,16 @@ def _extract_http_macaroon(request):
     return None
 
 
-@implementer(ISecurityPolicy)
-class MacaroonSecurityPolicy:
-    def identity(self, request):
+@implementer(IAuthenticationPolicy)
+class MacaroonAuthenticationPolicy(CallbackAuthenticationPolicy):
+    def __init__(self, callback=None):
+        self.callback = callback
+
+    def unauthenticated_userid(self, request):
         # If we're calling into this API on a request, then we want to register
         # a callback which will ensure that the response varies based on the
         # Authorization header.
         request.add_response_callback(add_vary_callback("Authorization"))
-        request.authentication_method = AuthenticationMethod.MACAROON
 
         # We need to extract our Macaroon from the request.
         macaroon = _extract_http_macaroon(request)
@@ -87,7 +89,9 @@ class MacaroonSecurityPolicy:
         # Check to see if our Macaroon exists in the database, and if so
         # fetch the user that is associated with it.
         macaroon_service = request.find_service(IMacaroonService, context=None)
-        return macaroon_service.find_from_raw(macaroon).user
+        userid = macaroon_service.find_userid(macaroon)
+        if userid is not None:
+            return str(userid)
 
     def remember(self, request, userid, **kw):
         # This is a NO-OP because our Macaroon header policy doesn't allow
@@ -95,19 +99,11 @@ class MacaroonSecurityPolicy:
         # assumes it has been configured in clients somewhere out of band.
         return []
 
-    def forget(self, request, **kw):
+    def forget(self, request):
         # This is a NO-OP because our Macaroon header policy doesn't allow
         # the ability for authentication to "forget" the user id. This
         # assumes it has been configured in clients somewhere out of band.
         return []
-
-    def authenticated_userid(self, request):
-        # Handled by MultiSecurityPolicy
-        return NotImplemented
-
-    def permits(self, request, context, permission):
-        # Handled by MultiSecurityPolicy
-        return NotImplemented
 
 
 @implementer(IAuthorizationPolicy)
